api/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
anthropic>=0.40.0
openai>=1.12.0
google-generativeai>=0.8.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
aiofiles>=23.2.0
sse-starlette>=1.8.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
httpx>=0.26.0
